requirements.txt
asn1crypto>=1.5.1
certifi>=2024.2.2
cffi>=1.16.0
chardet>=5.2.0
cryptography>=42.0.5
idna>=3.6
pycparser>=2.21
pyOpenSSL>=24.0.0
PySocks>=1.7.1
PyYAML>=6.0.1
requests>=2.31.0
six>=1.16.0
urllib3>=2.1.0